Fix the export for real this time
